fix with negative duration
Travis can send a negative duration from its api.
We need to investigate why, it seems to be a bug on their api.

+++ common/common_test.go
@@ -70,7 +70,7 @@ var _ = Describe("Common", func() {
 		Context("with build in success", func() {
 			build := travis.Build{
 				State: travis.SUCCEEDED_STATE,
-				Duration: uint(60),
+				Duration: 60,
 				StartedAt: "now",
 			}
 			expectedMetadata := []model.Metadata{
